app/api/requirements.txt
tensorflow-cpu>=2.15.0
fastapi>=0.103.1
uvicorn>=0.23.2
python-multipart>=0.0.6
Pillow>=10.0.0
numpy>=1.24.3
